requirements.txt
-r lib/mccole/requirements.txt
greenlet
kaleido
pandas
plotly
prettytable
py-des-lib
pydot
pyfakefs
pytest
requests
z3-solver
